refactor(api-security-user-management): new security models
